Switch back to only encoding once, and being sure we pass encoded values all the way through to backend
Also, make sure Pages are passed decoded arguments, so they don't have to take that responsibility.

The only remaining URLs that fail to work have encoded <, >, :, / or ? in them.  We're passing them to the
backend correctly, we think, but the backend web server is not interpreting them correctly.  Bugs filed.

This change depends heavily on the earlier fix that makes window.location.hash behave the same in Firefox
as it does in other browsers.  That had been the source of a lot of router problems, which were patched
over by double-encoding.  Now that the router is safe, we don't need the double-encoding.

Conceptually, we should only ever encode things once, rather than double encoding sometimes,
 then guessing at how many times to decode.